Added custom tag file

+++ src/main/java/com/myfp/fp/dao/postgres/ConnectionRequestDAOPstSQL.java
@@ -213,7 +213,7 @@ public class ConnectionRequestDAOPstSQL extends BaseDAOImpl implements Connectio
             user.setId(rs.getLong("user_id"));
             user.setEmail(rs.getString("email"));
             user.setPass(rs.getString("pass"));
-            user.setUserBalance(rs.getInt("user_balance"));
+            user.setUserBalance(rs.getDouble("user_balance"));
             user.setFirstname(rs.getString("firstname"));
             user.setMiddleName(rs.getString("middle_name"));
             user.setSurname(rs.getString("surname"));
